+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 TBD
 
+- Request auto refresh configurable through `ui.auto_refresh` config option.
 - Added additional matching for topic subscriptions to include full string, empty, None, and regular expressions.
 - Fixed Request system version to return actual system version instead of parsed version when using latest.
 - Expanded the Topics API to support `api/v1/topics/name/` for passing `topic.name` as a variable instead of `topic.id`